Handle undefined variables in rss templates
These templates used to work when using the StrictUndefined undefined
behavior in the jinja2 environment, but in recent releases, this causes
the build to fail. This issue can be seen at
https://github.com/Python-Community-News/Site/pull/38. Using the `is
defined` test results in similar behavior even when `StrictUndefined` is
set, allowing site developers to catch typos or other missing template
context without breaking their rss feeds.

This change also includes a new test and a test fixture to create a
jinja2 environment that can be modified with the scope of a single test.